+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 
 - [Use Bootstrap media breakpoints](https://github.com/catalpainternational/myeqip-portal/pull/137)
 - [Mobile widgets were hidden on Safari](https://github.com/catalpainternational/myeqip-portal/pull/142)
+- [Title block supports a banner image](https://github.com/catalpainternational/myeqip-portal/pull/144)
 
 ## 0.1.0